Mater user now can view all

+++ app/Http/Controllers/UserController.php
@@ -10,26 +10,32 @@ class UserController extends Controller
 {
   public function index()
   {
+    if(Auth::user()->role->name!='master') return response('you dont have access to this',400);
     return User::all();
   }
 
   public function store()
   {
+    if(Auth::user()->role->name!='master') return response('you dont have access to this',400);
     return User::create(request()->all());
   }
 
   public function get(User $user)
   {
+    if(Auth::user()->role->name!='master') return response('you dont have access to this',400);
     return $user;
   }
 
   public function actualUser()
   {
-    return Auth::user();
+    $user = Auth::user();
+    $user->role;
+    return $user;
   }
 
   public function update(User $user)
   {
+    if(Auth::user()->role->name!='master' && $user->id!=Auth::id()) return response('you dont have access to this',400);
     $user->fill(request()->all());
     $user->save();
     return $user;
@@ -37,22 +43,8 @@ class UserController extends Controller
 
   public function delete(User $user)
   {
+    if(Auth::user()->role->name!='master') return response('you dont have access to this',400);
     $user->delete();
     return 'user deleted';
-  }
-
-  public function role(User $user)
-  {
-    return $user->role;
-  }
-
-  public function address(User $user)
-  {
-    return $user->address;
-  }
-
-  public function clients(User $user)
-  {
-    return $user->clients;
   }
 }
